feat: integrate native TLOS OFT into main bridge interface
