+++ src/features/bridge/ui/GasOnDestinationButton.tsx
@@ -4,7 +4,6 @@ import {uiStore} from '@/core/stores/uiStore';
 import {styled} from '@/core/ui/system';
 
 import {bridgeStore} from '../stores/bridgeStore';
-import { nativeBridgeStore } from '../stores/nativeBridgeStore';
 
 export const GasOnDestinationButton = observer(() => {
   const {form, dstNativeAmount} = bridgeStore;
@@ -28,26 +27,3 @@ const Button = styled('span', {name: 'GasOnDestinationButton'})(({theme}) => ({
   cursor: 'pointer',
 }));
 
-// duplicated template to reference native OFT store
-
-export const NativeGasOnDestinationButton = observer(() => {
-  const {form, dstNativeAmount} = nativeBridgeStore;
-  return (
-    <>
-      {form.dstChainId ? (
-        <GasButton onClick={uiStore.oftDstNativeAmountModal.open}>
-          {dstNativeAmount?.equalTo(0) || !dstNativeAmount
-            ? 'Add'
-            : `${dstNativeAmount.toExact()} ${dstNativeAmount.currency.symbol}`}
-        </GasButton>
-      ) : (
-        '--'
-      )}
-    </>
-  );
-});
-
-const GasButton = styled('span', {name: 'NativeGasOnDestinationButton'})(({theme}) => ({
-  color: theme.palette.primary.main,
-  cursor: 'pointer',
-}));
