+++ src/pages/_app.tsx
@@ -10,7 +10,6 @@ import React, {useEffect} from 'react';
 
 import {AccountModal} from '@/bridge/ui/AccountModal';
 import {GasOnDestinationModal} from '@/bridge/ui/GasOnDestinationModal';
-import { NativeGasOnDestinationModal } from '@/bridge/ui/GasOnDestinationModal';
 import {JsonRpcProviderContext} from '@/core/context/JsonRpcProviderContext';
 import {THEMES, useUserThemePreference} from '@/core/hooks/useUserThemePreference';
 import {UserThemePreferenceProvider} from '@/core/hooks/useUserThemePreference';
@@ -61,10 +60,6 @@ const App = observer(({Component, pageProps}: AppPropsWithLayout) => {
             <GasOnDestinationModal
               open={uiStore.dstNativeAmountModal.value}
               onClose={uiStore.dstNativeAmountModal.close}
-            />
-            <NativeGasOnDestinationModal
-              open={uiStore.oftDstNativeAmountModal.value}
-              onClose={uiStore.oftDstNativeAmountModal.close}
             />
             <AccountModal open={walletModal.value} onClose={walletModal.close} />
           </ToastProvider>
